CONFIG update
This was a great day for the app

Made new method for cursor and the cursor to read out of a dictionary.

App.py now calls the cursor. App.py also now has the start of a good search query from python to SQL.

Spreadsheet @ 10/14/21 is the stackoverflow for next steps.